Fixed plot output location
Fixed a bug that must have been introduced in commit
a824338947fddc7605523e0d6c6e7fb095e02ecf. Because `output` is now a
keyword argument to `plotypus.process_star`, it is not being passed to
`lightcurve.plot_lightcurve` through `**ops`. This results in
`lightcurve.plot_lightcurve` using the default value of `output='.'`,
causing all plots to be output in the current directory. It now passes
this value properly, and the issue is resolved.