Moved plotting settings to matplotlibrc
All plotting settings have been moved to a single `matplotlibrc` file,
located in the root of the package.

One downside is that the `matplotlibrc` file must be in a location which
`matplotlib` looks in. A list of locations is given
[here](http://matplotlib.org/users/customizing.html).

Also, by making plot colors configurable, the `color` option to
`lightcurve.plot_lightcurve` is a bit scattered now. If you want black
and white plots, you can configure the `axes.color_cycle` setting to
just black, but the `color` option still exists to do things like change
the point types. Perhaps the keyword should be changed to `colorblind`,
and have the opposite effect. Colorblind modes for programs typically
don't change the colors, but make the symbols unambiguous when there is
no color, and that's exactly what the `color` option does right now.