Added skiprows option and fixed division by zero error.
There is now a command line option --skiprows N, or just skiprows=N for
the get_lightcurve_from_file function, which causes the first N rows in
the photometry tables to be skipped. Useful when reading files with
extra information at the top.

Also changed the default min-period argument to get_lightcurve from 0 to
0.2, as it was causing division by zero.