Thought I got multiple periods working but actually broke everything
The problem with approaching multiple periods by using a 2D array of
phases, where each row contains the same times phased to different
periods, is that when calling ``Predictor.fit(phases, mags)``, it is not
expecting ``phases`` to have the shape it has.

We will have to take a different approach. Instead, ``time`` should be
passed in place of ``phases``, and the periods will have to be passed in
as well, so that rephasing takes place within ``Fourier.design_matrix``.