Moved phasing into ``Fourier.design_matrix``
Everything looks like it should work for 1 or more periods, aside from
formatting the output, but instead the fit is always a constant
zero. Everything is zero!! Aaaurrghh!

TODO: Compute a design_matrix by hand, compare it to the one being
produced by the function. It looks right, but I haven't tested it like
that.