Imported matplotlib
I removed matplotlib from the imports in lightcurve.py, because a code
analysis tool told me it was unused. I had forgotten that plotypus.py
actually accesses lightcurve.matplotlib in order to provide the
matplotlibrc file, so it was impossible to plot anything without putting
that import back in.

Oops.