Replaced separate --shifts and --periods files with --parameters file
The functionality once provided by the --shifts and --periods options
has now been replaced by a unified --parameters file, as well as --shift
and --period options which only take numbers as arguments, not files.

The parameters file is read as a pandas table, and thus pandas is a new
dependency.