Laying foundation for separate mpl/tikz plotting engines
I have added a new CLI argument, `--plotting-engine={mpl,tikz}`, which
is then passed to the newly changed `plot_lightcurve` in the `engine`
keyword argument.

`plot_lightcurve` is now a wrapper for two different functions:
`plot_lightcurve_mpl` and `plot_lightcurve_tikz`. It merely acts on the
`engine` argument, calls one of the two, and returns what they return.

`plot_lightcurve_mpl` is now a carbon copy of what `plot_lightcurve`
once was. It still needs to be modified to return a
`matplotlib.pyplot.Figure` object, instead of
nothing. `plot_lightcurve_tikz` is just a placeholder for now. Both also
need to be changed to act on the `dict` returned by `get_lightcurve`,
instead of bare values.