Added ability to output periodogram to table
Now the option `--output-table-periodogram` allows you to output the
periodogram to a tabular file, in units of
`--output-periodogram-form={period,frequency}`. The option to output a
plot still needs to be added.

This required a bit of modification to the periodogram functions, so
that they now optionally return the entire periodogram, in addition to the
optimal period. This is not the case by default, as it is extra
computation, and it can be set with the `output_periodogram` keyword argument.