Fixed directory processing issue
As reported in #88, somehow I broke the feature which allows processing
entire directories, by providing only the name of the directory. I'm
still not entirely sure what the cause was, but it was fixed by
resolving the pathnames as early as possible. We were actually doing
some extra computation for the other input modes, which this fix
removes, so that's a plus.