Fixed array dimension error and period finding bug
I have corrected two bugs.

The first is described in part of issue #94. There was an array
dimension mismatch, that was caused by using masked arrays for some
values, and only the un-masked elements for other values. The solution
was to use masked arrays for both.

Another bug was uncovered while fixing that one. In `get_lightcurve`, I
had previously done away with the `_period` variable, and used `period`
both as the function parameter and as an internal variable. It turns out
that the distinction between `_period` and `period` actually had a
subtle importance, as `period` never changed, causing a certain
conditional to behave as expected. A new variable would need to be
introduced in order to do away with `_period`, so I changed things back
to their original form.