Fixed array index error caused by no outliers
I have fixed the error described in issue #94, caused by a lack of
outliers. The array index error is just a symptom, the real problem is
that `matplotlib.pyplot.errorbar` will try to access element `0` in the
`yerr` array, even if it is empty. This is a bug in matplotlib.

The solution was to replace `yerr` with `None` whenever the value being
passed had length zero.