Reducing unnecessary phasing
There is a lot of coupling when it comes to phased/unphased data, so I
had to make a lot of changes in this one commit.

The main point of this commit was to deal with temporal-space data
instead of phase-space data wherever possible. This was especially
important in the `Fourier` class, to prepare for multiperiodic signal
support. Now the period needs to be provided to `Fourier` before
`Fourier.fit` or `Fourier.transform` is called. To allow for setting the
`period` parameter properly when `Fourier` is nested within another
estimator, I had to make `Fourier` a subclass of scikit-learn's
`BaseEstimator`, which we should have done long ago.

Most of the changes in `get_lightcurve` deal with replacing `phase` with
`time`. There are also some changes which deal with providing the
correct data when dealing with masked arrays. There is also a new piece
of code which passes the period into the `Fourier` object, which
required a little bit of hackery (see the comments before the call to
`predictor.set_params`).

The `rephase` function in `plotypus.periodogram` was also modified, such
that masked values are still rephased.