Removed intercept from design matrix
Now the Fourier design matrix no longer provides the intercept column,
and relies on the regressor object to fit the intercept separately. This
is already the default behavior of all sklearn regressors (that I'm
aware of), and so this simply saves the user from needing to provide
`fit_intercept=False` every time.