Replaced hand-chosen regressors with completely arbitrary ones
Instead of having the `--regressor` option take one of a few hand-chosen
values, it now takes the name of _any arbitrary regressor object_, such
as the default, `--regressor sklearn.linear_model.LassoLarsIC`. This
requires a little more user-knowledge, but opens a world of
possibilities.

This is accomplished through the new `plotypus.utils.import_name`
function, which imports an object from its full name and returns it,
without affecting the enclosing namespace.