Generalized phasing within `lightcurve.plot_lightcurve`
A number of changes have been made relating to phasing and plotting, in
preparation for multiple periods.

Before, the `lightcurve.get_lightcurve` function returned an array
`phased_data`, which contained the original data, with outliers masked,
and with time replaced with phase. Now it has been replaced with the
array `data`, which only differs from the input by masking outliers.

The function also returned fit values with phase [0, 1). Now it instead
returns fit values over the entire time interval, and leaves phasing for
later. This does result in unnecessary computation for singly-periodic
stars, but it will be necessary for multi-periodic stars, and avoiding
the extra computation in the special case would require adding
complexity to the code. This array has been renamed from `lightcurve` to
`data_smooth`, to emphasize its relation to `data`.

Now, instead of plotting the data and fit repeated over 2 periods, we
phase the data such that it repeats after 2 periods instead of one. This
removes the need to duplicate the arrays, causes the observed data
points on each half to be different, and generalizes to multiple
periods, where the data _should not_ repeat after 1 period. This also
allows us to make the number of cycles before repetition _arbitrary_,
and thus a new CLI option `--cycles=2.0` has been added. `cycles` has
also been made a keyword argument to all functions which involve
rephasing, as well as the `plot_lightcurve` functions.

The option to plot _unphased_ data has also been created, although it is
currently inaccessible from the CLI. Simply plot data with
`period=None`, and it will be left unphased. This is rather ugly when
outputting to a small image, and will look much better if we eventually
add support for a format which supports interactive scrolling, or simply
use a matplotlibrc with a _very_ wide figure size.

The unused `--phase-points` CLI option has been removed.

A missing comma was added to the list given by the `__all__` attribute
of `plotypus.utils`.

Some doc strings are out of sync with the changes, though they were out
of sync before the changes as well.