Modifying lightcurve output table to output 2nd period
Previously we were outputting the fit over the duration of the first
period, which meant leaving out a number of phase points. Now we instead
output the second period, which is problematic in itself, but is at
least an improvement in most real scenarios. Ideally we should just be
outputting phase points which are not sensitive to any period.